rename Autosuggest to be deprecated
