--- src/components/Autosuggest/index.js
@@ -1,2 +1,0 @@
-export { Autosuggest } from './Autosuggest';
-export { ItemTag } from './ItemTag';
