+++ src/components/AutosuggestDeprecated/AutosuggestDeprecated.scss
@@ -1,4 +1,4 @@
-.Autosuggest {
+.AutosuggestDeprecated {
     position: relative;
     min-height: 38px;
