+++ src/components/AutosuggestDeprecated/__mocks__/suggestions.js
