+++ src/components/AutosuggestDeprecated/index.js
@@ -1,0 +1,2 @@
+export { AutosuggestDeprecated } from './AutosuggestDeprecated';
+export { ItemTag } from './ItemTag';
