+++ src/components/SelectComponents/AutosuggestMulti/__tests__/AutosuggestMulti.spec.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import toJson from 'enzyme-to-json';
 import { AutosuggestMulti } from '../AutosuggestMulti';
-import { SUGGESTIONS, SUGGESTION_TO_STRING } from '../../../Autosuggest/__mocks__/suggestions';
+import {
+    SUGGESTIONS,
+    SUGGESTION_TO_STRING,
+} from '../../../AutosuggestDeprecated/__mocks__/suggestions';
 
 describe('AutosuggestMulti', () => {
     const suggestionToString = SUGGESTION_TO_STRING;
