+++ src/components/SelectComponents/ComboboxMulti/__tests__/ComboboxMulti.spec.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import toJson from 'enzyme-to-json';
 import { ComboboxMulti } from '../ComboboxMulti';
-import { SUGGESTIONS, SUGGESTION_TO_STRING } from '../../../Autosuggest/__mocks__/suggestions';
+import {
+    SUGGESTIONS,
+    SUGGESTION_TO_STRING,
+} from '../../../AutosuggestDeprecated/__mocks__/suggestions';
 
 describe('ComboboxMulti', () => {
     const suggestionToString = SUGGESTION_TO_STRING;
