+++ src/components/SelectComponents/Select/__tests__/Select.spec.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import toJson from 'enzyme-to-json';
 import { Select } from '../Select';
-import { SUGGESTIONS, SUGGESTION_TO_STRING } from '../../../Autosuggest/__mocks__/suggestions';
+import {
+    SUGGESTIONS,
+    SUGGESTION_TO_STRING,
+} from '../../../AutosuggestDeprecated/__mocks__/suggestions';
 
 describe('Select', () => {
     const mockOnChange = jest.fn();
